renamed redirect component
